fix 39 bugs found during full codebase audit
Critical: dashboard optimization toggles broken (tunneling event blocked
CheckBox clicks), NvAPI QueryInterface returned int instead of IntPtr
(64-bit pointer truncation), update batch script path injection.

High: GameDetector crashed on built-in profile match (foreach mutation)
and had unsynchronized list access across threads. ProcessSnapshotService
stored dangling OS handles after Dispose. SystemStateSnapshot IntPtr
fields broke JSON serialization. HybridCpuDetector leaked process handles
every session. TrayIconManager.TogglePause deadlocked UI thread.
TimerResolutionService passed wrong value on stop. Deadlock game profile
used wrong process name. DPC Doctor rebooted without confirmation.
Dashboard event handlers accumulated on navigate-back.

Medium: 10 RunProcess helpers had stdout/stderr deadlock pattern.
PingMonitor race condition on Stop could crash process. JournalManager
and SettingsManager lacked thread safety. Directory prefix match was
too broad. DPC fix revert assumed DWORD type. Task deferral re-enabled
user-disabled tasks. Multiple event subscription leaks across ViewModels.
PowerPlanSwitcher activated plans as side effect of existence check.

+++ src/GameShift.Core/Optimization/Gpu/NvApiDrsManager.cs
@@ -72,7 +72,7 @@ public sealed class NvApiDrsManager : IDisposable
     // ── Delegates ─────────────────────────────────────────────────────────────
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-    private delegate int NvAPI_QueryInterface_t(uint functionId);
+    private delegate IntPtr NvAPI_QueryInterface_t(uint functionId);
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
     private delegate int NvAPI_Initialize_t();
@@ -405,7 +405,7 @@ public sealed class NvApiDrsManager : IDisposable
 
     private static T? GetDelegate<T>(NvAPI_QueryInterface_t qi, uint functionId) where T : Delegate
     {
-        var ptr = (IntPtr)qi(functionId);
+        var ptr = qi(functionId);
         return ptr == IntPtr.Zero ? null : Marshal.GetDelegateForFunctionPointer<T>(ptr);
     }
